Omicron, OpenCV, SmarActHCU compiler warning fixes
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17197 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Omicron/OmicronDeviceDriver.cpp
@@ -127,7 +127,7 @@ void checkforDevices(void) {
 		int nrDevices = 0;
 		ExXGetNumberOfDevices(&nrDevices);
 		if (nrDevices > alldevices.size()) {
-			int oldSize = alldevices.size();
+			int oldSize = (int) alldevices.size();
 			alldevices.resize(nrDevices);
 			for (int i = oldSize; i < nrDevices; i++) {
 				TxX_Error res = ExXGetDeviceID(i, &alldevices[i].devID);
